[FIX] mail_tracking: mixed classes
In migration #873 the classes DiscussController and MailController which
are the result of an upstream functionality split are merge altogether
in this module override causing undesired side effects in other modules
trying to extend those (i.e.: portal)

TT41432
